pyproject.toml
[project]
name = "youtubesynth"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=1.6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=1.0.0",
    "youtube-transcript-api>=0.6.0",
    "yt-dlp>=2024.0.0",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-mock>=3.12",
    "pytest-cov>=4.0",
]

[project.scripts]
youtubesynth = "youtubesynth.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["youtubesynth*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
